adding auth middleware

+++ controllers/sauce.js
@@ -30,7 +30,21 @@ exports.modifySauce = (req, res, next) => {
 }
 
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.params.id })
-    .then(sauce => res.status(200).json({ message: 'Objet supprimé !' }))
-    .catch(error => res.status(400).json({ error }));
+  Sauce.findOne({ _id: req.params.id }).then(
+    (sauce) => {
+      if (!sauce) {
+        return res.status(404).json({
+          error: new Error('Objet non trouvé !')
+        });
+      }
+      if (sauce.userId !== req.auth.userId) {
+        res.status(401).json({
+          error: new Error('Requête non authorisée !')
+        });
+      Sauce.deleteOne({ _id: req.params.id })
+        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
+        .catch(error => res.status(400).json({ error }));
+      }
+    }
+  );
 }
